Split RegPolyTool, SwitchTool and UserSVGTool
