+++ srcjs/swiss-army-knife-card.js
@@ -62,6 +62,9 @@ import CircularSliderTool from './circularslidertool';
 import RangeSliderTool from './rangeslidertool';
 import CircleTool from './circletool';
 import LineTool from './linetool';
+import SwitchTool from './switchtool';
+import RegPolyTool from './regpolytool';
+import UserSvgTool from './usersvg';
 
 // Original injector is buggy. Use a patched version, and store this local...
 import * as SvgInjector from '../dist/SVGInjector.min.js'; // lgtm[js/unused-local-variable]
@@ -432,601 +435,15 @@ class Toolset {
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
 
-/** ****************************************************************************
-  * SwitchTool class
-  *
-  * Summary.
-  *
-  *
-  * NTS:
-  * - .mdc-switch__native-control uses:
-  *     - width: 68px, 17em
-  *     - height: 48px, 12em
-  * - and if checked (.mdc-switch--checked):
-  *     - transform: translateX(-20px)
-  *
-  * .mdc-switch.mdc-switch--checked .mdc-switch__thumb {
-  *  background-color: var(--switch-checked-button-color);
-  *  border-color: var(--switch-checked-button-color);
-  *
-  */
-
-class SwitchTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_SWITCH_CONFIG = {
-      position: {
-        cx: 50,
-        cy: 50,
-        orientation: 'horizontal',
-        track: {
-          width: 16,
-          height: 7,
-          radius: 3.5,
-        },
-        thumb: {
-          width: 9,
-          height: 9,
-          radius: 4.5,
-          offset: 4.5,
-        },
-      },
-      classes: {
-        tool: {
-          'sak-switch': true,
-          hover: true,
-        },
-        track: {
-          'sak-switch__track': true,
-        },
-        thumb: {
-          'sak-switch__thumb': true,
-        },
-      },
-      styles: {
-        tool: {
-        },
-        track: {
-        },
-        thumb: {
-        },
-      },
-    };
-
-    const HORIZONTAL_SWITCH_CONFIG = {
-      animations: [
-        {
-          state: 'on',
-          id: 1,
-          styles: {
-            track: {
-              fill: 'var(--switch-checked-track-color)',
-              'pointer-events': 'auto',
-            },
-            thumb: {
-              fill: 'var(--switch-checked-button-color)',
-              transform: 'translateX(4.5em)',
-              'pointer-events': 'auto',
-            },
-          },
-        },
-        {
-          state: 'off',
-          id: 0,
-          styles: {
-            track: {
-              fill: 'var(--switch-unchecked-track-color)',
-              'pointer-events': 'auto',
-            },
-            thumb: {
-              fill: 'var(--switch-unchecked-button-color)',
-              transform: 'translateX(-4.5em)',
-              'pointer-events': 'auto',
-            },
-          },
-        },
-      ],
-    };
-
-    const VERTICAL_SWITCH_CONFIG = {
-      animations: [
-        {
-          state: 'on',
-          id: 1,
-          styles: {
-            track: {
-              fill: 'var(--switch-checked-track-color)',
-              'pointer-events': 'auto',
-            },
-            thumb: {
-              fill: 'var(--switch-checked-button-color)',
-              transform: 'translateY(-4.5em)',
-              'pointer-events': 'auto',
-            },
-          },
-        },
-        {
-          state: 'off',
-          id: 0,
-          styles: {
-            track: {
-              fill: 'var(--switch-unchecked-track-color)',
-              'pointer-events': 'auto',
-            },
-            thumb: {
-              fill: 'var(--switch-unchecked-button-color)',
-              transform: 'translateY(4.5em)',
-              'pointer-events': 'auto',
-            },
-          },
-        },
-      ],
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_SWITCH_CONFIG, argConfig), argPos);
-
-    if (!['horizontal', 'vertical'].includes(this.config.position.orientation))
-      throw Error('SwitchTool::constructor - invalid orientation [vertical, horizontal] = ', this.config.position.orientation);
-
-    this.svg.track = {};
-    this.svg.track.radius = Utils.calculateSvgDimension(this.config.position.track.radius);
-
-    this.svg.thumb = {};
-    this.svg.thumb.radius = Utils.calculateSvgDimension(this.config.position.thumb.radius);
-    this.svg.thumb.offset = Utils.calculateSvgDimension(this.config.position.thumb.offset);
-
-    switch (this.config.position.orientation) {
-      default:
-      case 'horizontal':
-        this.config = Merge.mergeDeep(DEFAULT_SWITCH_CONFIG, HORIZONTAL_SWITCH_CONFIG, argConfig);
-
-        this.svg.track.width = Utils.calculateSvgDimension(this.config.position.track.width);
-        this.svg.track.height = Utils.calculateSvgDimension(this.config.position.track.height);
-        this.svg.thumb.width = Utils.calculateSvgDimension(this.config.position.thumb.width);
-        this.svg.thumb.height = Utils.calculateSvgDimension(this.config.position.thumb.height);
-
-        this.svg.track.x1 = this.svg.cx - this.svg.track.width / 2;
-        this.svg.track.y1 = this.svg.cy - this.svg.track.height / 2;
-
-        this.svg.thumb.x1 = this.svg.cx - this.svg.thumb.width / 2;
-        this.svg.thumb.y1 = this.svg.cy - this.svg.thumb.height / 2;
-        break;
-
-      case 'vertical':
-        this.config = Merge.mergeDeep(DEFAULT_SWITCH_CONFIG, VERTICAL_SWITCH_CONFIG, argConfig);
-
-        this.svg.track.width = Utils.calculateSvgDimension(this.config.position.track.height);
-        this.svg.track.height = Utils.calculateSvgDimension(this.config.position.track.width);
-        this.svg.thumb.width = Utils.calculateSvgDimension(this.config.position.thumb.height);
-        this.svg.thumb.height = Utils.calculateSvgDimension(this.config.position.thumb.width);
-
-        this.svg.track.x1 = this.svg.cx - this.svg.track.width / 2;
-        this.svg.track.y1 = this.svg.cy - this.svg.track.height / 2;
-
-        this.svg.thumb.x1 = this.svg.cx - this.svg.thumb.width / 2;
-        this.svg.thumb.y1 = this.svg.cy - this.svg.thumb.height / 2;
-        break;
-    }
-
-    this.classes.track = {};
-    this.classes.thumb = {};
-
-    this.styles.track = {};
-    this.styles.thumb = {};
-    if (this.dev.debug) console.log('SwitchTool constructor config, svg', this.toolId, this.config, this.svg);
-  }
-
-  /** *****************************************************************************
-  * SwitchTool::value()
-  *
-  * Summary.
-  * Receive new state data for the entity this switch is linked to. Called from set hass;
-  *
-  */
-  set value(state) {
-    const changed = super.value = state;
-
-    return changed;
-  }
-
-  /**
-  * SwitchTool::_renderSwitch()
-  *
-  * Summary.
-  * Renders the switch using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the switch
-  *
-  */
-
-  _renderSwitch() {
-    this.MergeAnimationClassIfChanged();
-    // this.MergeColorFromState(this.styles);
-    this.MergeAnimationStyleIfChanged(this.styles);
-    // this.MergeAnimationStyleIfChanged(this.styles.thumb);
-
-    return svg`
-      <g>
-        <rect class="${classMap(this.classes.track)}" x="${this.svg.track.x1}" y="${this.svg.track.y1}"
-          width="${this.svg.track.width}" height="${this.svg.track.height}" rx="${this.svg.track.radius}"
-          style="${styleMap(this.styles.track)}"
-        />
-        <rect class="${classMap(this.classes.thumb)}" x="${this.svg.thumb.x1}" y="${this.svg.thumb.y1}"
-          width="${this.svg.thumb.width}" height="${this.svg.thumb.height}" rx="${this.svg.thumb.radius}" 
-          style="${styleMap(this.styles.thumb)}"
-        />
-      </g>
-      `;
-  }
-
-  /** *****************************************************************************
-  * SwitchTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  * https://codepen.io/joegaffey/pen/vrVZaN
-  *
-  */
-
-  render() {
-    return svg`
-      <g id="switch-${this.toolId}" class="${classMap(this.classes.tool)}" overflow="visible" transform-origin="${this.svg.cx} ${this.svg.cy}"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderSwitch()}
-      </g>
-    `;
-  }
-} // END of class
-
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
 
-/** ****************************************************************************
-  * RegPolyTool class
-  *
-  * Summary.
-  *
-  */
-
-class RegPolyTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_REGPOLY_CONFIG = {
-      position: {
-        cx: 50,
-        cy: 50,
-        radius: 50,
-        side_count: 6,
-        side_skip: 1,
-        angle_offset: 0,
-      },
-      classes: {
-        tool: {
-          'sak-polygon': true,
-          hover: true,
-        },
-        regpoly: {
-          'sak-polygon__regpoly': true,
-        },
-      },
-      styles: {
-        tool: {
-        },
-        regpoly: {
-        },
-      },
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_REGPOLY_CONFIG, argConfig), argPos);
-
-    this.svg.radius = Utils.calculateSvgDimension(argConfig.position.radius);
-
-    this.classes.regpoly = {};
-    this.styles.regpoly = {};
-    if (this.dev.debug) console.log('RegPolyTool constructor config, svg', this.toolId, this.config, this.svg);
-  }
-
-  /** *****************************************************************************
-  * RegPolyTool::value()
-  *
-  * Summary.
-  * Receive new state data for the entity this circle is linked to. Called from set hass;
-  *
-  */
-  set value(state) {
-    const changed = super.value = state;
-
-    return changed;
-  }
-
-  /** *****************************************************************************
-  * RegPolyTool::_renderRegPoly()
-  *
-  * Summary.
-  * Renders the regular polygon using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the regular polygon
-  *
-  */
-
-  _renderRegPoly() {
-    const generatePoly = function (p, q, r, a, cx, cy) {
-      const base_angle = 2 * Math.PI / p;
-      let angle = a + base_angle;
-      let x; let y; let
-        d_attr = '';
-
-      for (let i = 0; i < p; i++) {
-        angle += q * base_angle;
-
-        // Use ~~ as it is faster then Math.floor()
-        x = cx + ~~(r * Math.cos(angle));
-        y = cy + ~~(r * Math.sin(angle));
-
-        d_attr
-          += `${((i === 0) ? 'M' : 'L') + x} ${y} `;
-
-        if (i * q % p === 0 && i > 0) {
-          angle += base_angle;
-          x = cx + ~~(r * Math.cos(angle));
-          y = cy + ~~(r * Math.sin(angle));
-
-          d_attr += `M${x} ${y} `;
-        }
-      }
-
-      d_attr += 'z';
-      return d_attr;
-    };
-
-    this.MergeAnimationStyleIfChanged();
-    this.MergeColorFromState(this.styles.regpoly);
-
-    return svg`
-      <path class="${classMap(this.classes.regpoly)}"
-        d="${generatePoly(this.config.position.side_count, this.config.position.side_skip, this.svg.radius, this.config.position.angle_offset, this.svg.cx, this.svg.cy)}"
-        style="${styleMap(this.styles.regpoly)}"
-      />
-      `;
-  }
-
-  /** *****************************************************************************
-  * RegPolyTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  //        @click=${e => this._card.handlePopup(e, this._card.entities[this.defaultEntityIndex()])} >
-
-  render() {
-    return svg`
-      <g "" id="regpoly-${this.toolId}" class="${classMap(this.classes.tool)}" transform-origin="${this.svg.cx} ${this.svg.cy}"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderRegPoly()}
-      </g>
-    `;
-  }
-} // END of class
-
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
-
-/** ****************************************************************************
-  * UserSvgTool class, UserSvgTool::constructor
-  *
-  * Summary.
-  *
-  */
-
-class UserSvgTool extends BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    const DEFAULT_USERSVG_CONFIG = {
-      position: {
-        cx: 50,
-        cy: 50,
-        height: 50,
-        width: 50,
-      },
-      styles: {
-        usersvg: {
-        },
-        mask: {
-          fill: 'white',
-        },
-      },
-    };
-
-    super(argToolset, Merge.mergeDeep(DEFAULT_USERSVG_CONFIG, argConfig), argPos);
-
-    this.images = {};
-    this.images = Object.assign({}, ...this.config.images);
-
-    // #TODO:
-    // Select first key in k/v store. HOw??
-    this.item = {};
-    this.item.image = 'default';
-
-    // https://github.com/flobacher/SVGInjector2
-    // Note: in defs, url from gradient is changed, but NOT in the SVG fill=...
-
-    this.injector = {};
-    // Options
-    this.injector.injectorOptions = {
-      evalScripts: 'once',
-      pngFallback: 'assets/png',
-    };
-
-    this.injector.afterAllInjectionsFinishedCallback = function (totalSVGsInjected) {
-      // Callback after all SVGs are injected
-      // console.log('We injected ' + totalSVGsInjected + ' SVG(s)!');
-    };
-
-    this.injector.perInjectionCallback = function (svg) {
-      // Callback after each SVG is injected
-      this.injector.svg = svg;
-      // console.log('SVG injected: ', svg, this.injector);
-    }.bind(this);
-
-    // create injector configured by options
-    // this.injector.injector = new SVGInjector(this.injector.injectorOptions);
-
-    this.clipPath = {};
-
-    if (this.config.clip_path) {
-      this.svg.cp_cx = Utils.calculateSvgCoordinate(this.config.clip_path.position.cx || this.config.position.cx, 0);
-      this.svg.cp_cy = Utils.calculateSvgCoordinate(this.config.clip_path.position.cy || this.config.position.cy, 0);
-      this.svg.cp_height = Utils.calculateSvgDimension(this.config.clip_path.position.height || this.config.position.height);
-      this.svg.cp_width = Utils.calculateSvgDimension(this.config.clip_path.position.width || this.config.position.width);
-
-      const maxRadius = Math.min(this.svg.cp_height, this.svg.cp_width) / 2;
-
-      this.svg.radiusTopLeft = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-        this.config.clip_path.position.radius.top_left || this.config.clip_path.position.radius.left
-                                || this.config.clip_path.position.radius.top || this.config.clip_path.position.radius.all,
-      ))) || 0;
-
-      this.svg.radiusTopRight = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-        this.config.clip_path.position.radius.top_right || this.config.clip_path.position.radius.right
-                                || this.config.clip_path.position.radius.top || this.config.clip_path.position.radius.all,
-      ))) || 0;
-
-      this.svg.radiusBottomLeft = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-        this.config.clip_path.position.radius.bottom_left || this.config.clip_path.position.radius.left
-                                || this.config.clip_path.position.radius.bottom || this.config.clip_path.position.radius.all,
-      ))) || 0;
-
-      this.svg.radiusBottomRight = +Math.min(maxRadius, Math.max(0, Utils.calculateSvgDimension(
-        this.config.clip_path.position.radius.bottom_right || this.config.clip_path.position.radius.right
-                                || this.config.clip_path.position.radius.bottom || this.config.clip_path.position.radius.all,
-      ))) || 0;
-    }
-
-    if (this.dev.debug) console.log('UserSvgTool constructor config, svg', this.toolId, this.config, this.svg);
-  }
-
-  /** *****************************************************************************
-  * UserSvgTool::value()
-  *
-  * Summary.
-  * Receive new state data for the entity this usersvg is linked to. Called from set hass;
-  *
-  */
-  set value(state) {
-    const changed = super.value = state;
-
-    return changed;
-  }
-
-  updated(changedProperties) {
-    this.injector.elementsToInject = this._card.shadowRoot.querySelectorAll('svg[data-src]');
-    // console.log("updated - ", this._card.shadowRoot.getElementById("usersvg-".concat(this.toolId)), this.injector.elementsToInject);
-
-    this.injector.elementsToInject = this._card.shadowRoot.getElementById('usersvg-'.concat(this.toolId)).querySelectorAll('svg[data-src]:not(.injected-svg)');
-
-    // Trigger the injection if there is something to inject...
-    if (this.injector.elementsToInject.length > 0)
-      this.injector.injector.inject(
-        this.injector.elementsToInject,
-        this.injector.afterAllInjectionsFinishedCallback,
-        this.injector.perInjectionCallback,
-      );
-  }
-
-  /** *****************************************************************************
-  * UserSvgTool::_renderUserSvg()
-  *
-  * Summary.
-  * Renders the usersvg using precalculated coordinates and dimensions.
-  * Only the runtime style is calculated before rendering the usersvg
-  *
-  */
-
-  _renderUserSvg() {
-    this.MergeAnimationStyleIfChanged();
-
-    const images = Templates.getJsTemplateOrValue(this, this._stateValue, Merge.mergeDeep(this.images));
-
-    // if ((this.injector.svg) && (this.injector.image2.trim() === images[this.item.image].trim())) {
-    // return svg`${this.injector.svg}`;
-    // if (false) {
-    // } else {
-    if (images[this.item.image] === 'none')
-      return svg``;
-
-    let clipPath = '';
-    if (this.config.clip_path) {
-      clipPath = svg`
-        <defs>
-          <path  id="path-${this.toolId}"
-            d="
-              M ${this.svg.cp_cx + this.svg.radiusTopLeft + ((this.svg.width - this.svg.cp_width) / 2)} ${this.svg.cp_cy + ((this.svg.height - this.svg.cp_height) / 2)}
-              h ${this.svg.cp_width - this.svg.radiusTopLeft - this.svg.radiusTopRight}
-              a ${this.svg.radiusTopRight} ${this.svg.radiusTopRight} 0 0 1 ${this.svg.radiusTopRight} ${this.svg.radiusTopRight}
-              v ${this.svg.cp_height - this.svg.radiusTopRight - this.svg.radiusBottomRight}
-              a ${this.svg.radiusBottomRight} ${this.svg.radiusBottomRight} 0 0 1 -${this.svg.radiusBottomRight} ${this.svg.radiusBottomRight}
-              h -${this.svg.cp_width - this.svg.radiusBottomRight - this.svg.radiusBottomLeft}
-              a ${this.svg.radiusBottomLeft} ${this.svg.radiusBottomLeft} 0 0 1 -${this.svg.radiusBottomLeft} -${this.svg.radiusBottomLeft}
-              v -${this.svg.cp_height - this.svg.radiusBottomLeft - this.svg.radiusTopLeft}
-              a ${this.svg.radiusTopLeft} ${this.svg.radiusTopLeft}  0 0 1 ${this.svg.radiusTopLeft} -${this.svg.radiusTopLeft}
-              ">
-          </path>
-          <clipPath id="clip-path-${this.toolId}">
-            <use href="#path-${this.toolId}"/>
-          </clipPath>
-          <mask id="mask-${this.toolId}">
-            <use href="#path-${this.toolId}" style="${styleMap(this.styles.mask)}"/>
-          </mask>
-        </defs>
-        `;
-    }
-
-    // If svg, use injector for rendering. If jpg or png, use default image renderer...
-    if (['png', 'jpg'].includes((images[this.item.image].substring(images[this.item.image].lastIndexOf('.') + 1)))) {
-      // Render jpg or png
-      return svg`
-        <svg class="sak-usersvg__image" x="${this.svg.x}" y="${this.svg.y}" style="${styleMap(this.styles)}">
-          "${clipPath}"
-          <image clip-path="url(#clip-path-${this.toolId})" mask="url(#mask-${this.toolId})" href="${images[this.item.image]}" height="${this.svg.height}" width="${this.svg.width}"/>
-        </svg>
-        `;
-    } else {
-      return svg`
-        <svg class="sak-usersvg__image" data-some="${images[this.item.image]}" x="${this.svg.x}" y="${this.svg.y}" style="${styleMap(this.styles)}">
-          "${clipPath}"
-          <image clip-path="url(#clip-path-${this.toolId})" mask="url(#mask-${this.toolId})" href="${images[this.item.image]}" height="${this.svg.height}" width="${this.svg.width}"/>
-        </svg>
-        `;
-
-      // It seems new stuff is NOT injected for some reason. Donno why. Cant find it. Simply NOT injected, although injector is called in updated...
-      // 2022.07.24 For now, disable injector stuff...
-      // return svg`
-      // <svg id="image-one" data-src="${images[this.item.image]}" class="sak-usersvg__image" x="${this.svg.x}" y="${this.svg.y}"
-      // style="${styleMap(this.styles.usersvg)}" height="${this.svg.height}" width="${this.svg.width}">
-      // </svg>
-      // `;
-    }
-    // }
-  }
-
-  /** *****************************************************************************
-  * UserSvgTool::render()
-  *
-  * Summary.
-  * The render() function for this object.
-  *
-  */
-  render() {
-    return svg`
-      <g id="usersvg-${this.toolId}" overflow="visible" transform-origin="${this.svg.cx} ${this.svg.cy}"
-        style="${styleMap(this.styles.tool)}"
-        @click=${(e) => this.handleTapEvent(e, this.config)}>
-        ${this._renderUserSvg()}
-      </g>
-    `;
-  }
-} // END of class
 
 /**
   ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
